Add "format" field to attributes, add ForagePanTextUtils, add comments.
