+++ forage-android/src/main/java/com/joinforage/forage/android/ui/FormatPanTextWatcher.kt
@@ -1,10 +1,7 @@
 package com.joinforage.forage.android.ui
 
-import android.content.Context
-import android.content.res.TypedArray
 import android.text.Editable
 import android.text.TextWatcher
-import android.util.TypedValue
 import android.widget.EditText
 import com.joinforage.forage.android.model.queryForStateIIN
 
@@ -80,36 +77,6 @@ fun getTransformedPosition(pos: Int, oldPrettyText: String, newPrettyText: Strin
     // ensure we are within the newPrettyText's bounds
     val shiftedPos = pos + spacesShift
     return minOf(shiftedPos, newPrettyText.length)
-}
-
-fun getThemeAccentColor(context: Context): Int {
-    val outValue = TypedValue()
-    context.theme.resolveAttribute(android.R.attr.colorAccent, outValue, true)
-    return outValue.data
-}
-
-fun TypedArray.getBoxCornerRadiusBottomStart(boxCornerRadius: Float): Float {
-    val boxCornerRadiusBottomStart =
-        getDimension(com.joinforage.forage.android.R.styleable.ForagePANEditText_boxCornerBottomStart, 0f)
-    return if (boxCornerRadiusBottomStart == 0f) boxCornerRadius else boxCornerRadiusBottomStart
-}
-
-fun TypedArray.getBoxCornerRadiusTopEnd(boxCornerRadius: Float): Float {
-    val boxCornerRadiusTopEnd =
-        getDimension(com.joinforage.forage.android.R.styleable.ForagePANEditText_boxCornerTopEnd, 0f)
-    return if (boxCornerRadiusTopEnd == 0f) boxCornerRadius else boxCornerRadiusTopEnd
-}
-
-fun TypedArray.getBoxCornerRadiusBottomEnd(boxCornerRadius: Float): Float {
-    val boxCornerRadiusBottomEnd =
-        getDimension(com.joinforage.forage.android.R.styleable.ForagePANEditText_boxCornerBottomEnd, 0f)
-    return if (boxCornerRadiusBottomEnd == 0f) boxCornerRadius else boxCornerRadiusBottomEnd
-}
-
-fun TypedArray.getBoxCornerRadiusTopStart(boxCornerRadius: Float): Float {
-    val boxCornerRadiusTopStart =
-        getDimension(com.joinforage.forage.android.R.styleable.ForagePANEditText_boxCornerTopStart, 0f)
-    return if (boxCornerRadiusTopStart == 0f) boxCornerRadius else boxCornerRadiusTopStart
 }
 
 class FormatPanTextWatcher(
