+++ forage-android/src/main/res/values/attrs.xml
@@ -13,15 +13,15 @@
         <!-- Reusing android:textColor attribute -->
         <attr name="android:textSize" />
 
-        <attr name="cornerRadius" />
+        <attr format="dimension" name="cornerRadius" />
         <attr format="dimension" name="boxCornerTopEnd" />
         <attr format="dimension" name="boxCornerTopStart" />
         <attr format="dimension" name="boxCornerBottomStart" />
         <attr format="dimension" name="boxCornerBottomEnd" />
 
-        <attr name="android:strokeColor" />
-        <attr name="android:strokeWidth" />
-        <attr name="boxStrokeWidthFocused" />
+        <attr format="color" name="android:strokeColor" />
+        <attr format="float" name="android:strokeWidth" />
+        <attr format="dimension" name="boxStrokeWidthFocused" />
     </declare-styleable>
 
     <!-- Style attributes for the ForagePanEditText component. -->
