+++ sample-app/src/main/res/values/styles.xml
@@ -15,6 +15,7 @@
     <style name="TokenizeForageTextInputLayoutStyle" parent="Widget.MaterialComponents.TextInputLayout.OutlinedBox">
         <item name="android:hint">@string/tokenize_forage_edit_text_hint</item>
         <item name="boxStrokeWidth">1dp</item>
+        <item name="boxStrokeColor">@color/mtrl_textinput_default_box_stroke_color</item>
     </style>
 
     <!-- Style for the first ForagePanEditText in catalog Fragment. -->
